(PC-33936)[PRO] chore: remove suggested subcat feature, front-end side ☠
