+++ pro/src/pages/IndividualOffer/IndividualOfferDetails/commons/utils.ts
@@ -320,7 +320,6 @@ export function serializeDetailsPostData(
     durationMinutes: serializeDurationMinutes(formValues.durationMinutes ?? ''),
     extraData: serializeExtraData(formValues),
     productId: formValues.productId ? Number(formValues.productId) : undefined,
-    callId: formValues.callId,
     url: formValues.url,
   })
 }
