+++ pro/src/pages/IndividualOffer/IndividualOfferDetails/components/DetailsForm/DetailsForm.tsx
@@ -1,18 +1,13 @@
 import { useFormikContext } from 'formik'
-import React, { useState } from 'react'
-import { useDebouncedCallback } from 'use-debounce'
 
-import { api } from 'apiClient/api'
 import {
   CategoryResponseModel,
   SubcategoryResponseModel,
-  SuggestedSubcategoriesResponseModel,
 } from 'apiClient/v1'
 import { useIndividualOfferContext } from 'commons/context/IndividualOfferContext/IndividualOfferContext'
 import { CATEGORY_STATUS } from 'commons/core/Offers/constants'
 import { IndividualOfferImage } from 'commons/core/Offers/types'
 import { useActiveFeature } from 'commons/hooks/useActiveFeature'
-import { useSuggestedSubcategoriesAbTest } from 'commons/hooks/useSuggestedSubcategoriesAbTest'
 import { FormLayout } from 'components/FormLayout/FormLayout'
 import { OnImageUploadArgs } from 'components/ImageUploader/components/ButtonImageEdit/ModalImageEdit/ModalImageEdit'
 import fullMoreIcon from 'icons/full-more.svg'
@@ -28,9 +23,6 @@ import { InfoBox } from 'ui-kit/InfoBox/InfoBox'
 
 import { DetailsSubForm } from './DetailsSubForm/DetailsSubForm'
 import { Subcategories } from './Subcategories/Subcategories'
-import { SuggestedSubcategories } from './SuggestedSubcategories/SuggestedSubcategories'
-
-const DEBOUNCE_TIME_BEFORE_REQUEST = 400
 
 type DetailsFormProps = {
   isEanSearchDisplayed: boolean
@@ -57,82 +49,15 @@ export const DetailsForm = ({
   imageOffer,
   categoryStatus,
 }: DetailsFormProps): JSX.Element => {
-  const areSuggestedSubcategoriesUsed = useSuggestedSubcategoriesAbTest()
-  const [hasSuggestionsApiBeenCalled, setHasSuggestionsApiBeenCalled] =
-    useState(false)
-  const [suggestedSubcategories, setSuggestedSubcategories] = useState<
-    SuggestedSubcategoriesResponseModel['subcategoryIds']
-  >([])
-  const { values, setValues, handleChange, setFieldValue } =
-    useFormikContext<DetailsFormValues>()
-  const { subcategoryId, description, venueId, name, suggestedSubcategory } =
-    values
-  const { offer, subCategories } = useIndividualOfferContext()
+  const { values, handleChange } = useFormikContext<DetailsFormValues>()
+  const { subcategoryId } = values
+  const { offer } = useIndividualOfferContext()
   const offerAddressEnabled = useActiveFeature('WIP_ENABLE_OFFER_ADDRESS')
-
-  async function getSuggestedSubcategories() {
-    if (!areSuggestedSubcategoriesUsed && !offer) {
-      return
-    }
-
-    try {
-      const response = await api.getSuggestedSubcategories(
-        name,
-        description,
-        Number(venueId)
-      )
-
-      if (response.subcategoryIds.length === 0) {
-        throw new Error('No suggested subcategories')
-      }
-
-      await setFieldValue('callId', response.callId)
-      setSuggestedSubcategories(response.subcategoryIds)
-    } catch {
-      if (!suggestedSubcategory) {
-        await setValues({
-          ...values,
-          suggestedSubcategory: 'OTHER',
-        })
-      }
-    }
-
-    if (!hasSuggestionsApiBeenCalled) {
-      setHasSuggestionsApiBeenCalled(true)
-    }
-  }
-
-  const debouncedOnChangeGetSuggestedSubcategories = useDebouncedCallback(
-    getSuggestedSubcategories,
-    DEBOUNCE_TIME_BEFORE_REQUEST
-  )
-
-  async function onChangeGetSuggestedSubcategories(
-    e: React.ChangeEvent<
-      HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement
-    >
-  ) {
-    handleChange(e)
-    if (areSuggestedSubcategoriesUsed) {
-      await debouncedOnChangeGetSuggestedSubcategories()
-    }
-  }
-
-  const subcategory = subCategories.find((s) => s.id === subcategoryId)
 
   const isSubCategorySelected =
     subcategoryId !== DEFAULT_DETAILS_FORM_VALUES.subcategoryId
 
-  const showAddVenueBanner =
-    !areSuggestedSubcategoriesUsed && venuesOptions.length === 0
-
-  const isSuggestedSubcategoryDisplayed =
-    areSuggestedSubcategoriesUsed && !offer && !isProductBased
-
-  const showOtherAddVenueBanner =
-    areSuggestedSubcategoriesUsed &&
-    venuesOptions.length === 0 &&
-    subcategory?.onlineOfflinePlatform === CATEGORY_STATUS.OFFLINE
+  const showAddVenueBanner = venuesOptions.length === 0
 
   return (
     <>
@@ -170,20 +95,17 @@ export const DetailsForm = ({
                       ? 'Sélectionner la structure'
                       : 'Sélectionner le partenaire',
                   }}
-                  onChange={async (ev) => {
+                  onChange={(ev) => {
                     if (isProductBased) {
                       return
                     }
 
-                    await onChangeGetSuggestedSubcategories(ev)
+                    handleChange(ev)
                   }}
                   disabled={
                     readOnlyFields.includes('venueId') ||
                     venuesOptions.length === 1
                   }
-                  {...(isSuggestedSubcategoryDisplayed && {
-                    'aria-controls': 'suggested-subcategories',
-                  })}
                 />
               </FormLayout.Row>
             )}
@@ -193,11 +115,8 @@ export const DetailsForm = ({
                 label="Titre de l’offre"
                 maxLength={90}
                 name="name"
-                onChange={onChangeGetSuggestedSubcategories}
+                onChange={handleChange}
                 disabled={readOnlyFields.includes('name')}
-                {...(isSuggestedSubcategoryDisplayed && {
-                  'aria-controls': 'suggested-subcategories',
-                })}
               />
             </FormLayout.Row>
             <FormLayout.Row>
@@ -206,11 +125,8 @@ export const DetailsForm = ({
                 label="Description"
                 maxLength={1000}
                 name="description"
-                onChange={onChangeGetSuggestedSubcategories}
+                onChange={handleChange}
                 disabled={readOnlyFields.includes('description')}
-                {...(isSuggestedSubcategoryDisplayed && {
-                  'aria-controls': 'suggested-subcategories',
-                })}
               />
             </FormLayout.Row>
             {(categoryStatus === CATEGORY_STATUS.ONLINE ||
@@ -235,55 +151,23 @@ export const DetailsForm = ({
           </>
         )}
       </FormLayout.Section>
-      {isSuggestedSubcategoryDisplayed ? (
-        <SuggestedSubcategories
-          hasApiBeenCalled={hasSuggestionsApiBeenCalled}
-          suggestedSubcategories={suggestedSubcategories}
+      {!showAddVenueBanner && (
+        <Subcategories
           readOnlyFields={readOnlyFields}
           filteredCategories={filteredCategories}
           filteredSubcategories={filteredSubcategories}
         />
-      ) : (
-        !showAddVenueBanner && (
-          <Subcategories
-            readOnlyFields={readOnlyFields}
-            filteredCategories={filteredCategories}
-            filteredSubcategories={filteredSubcategories}
-          />
-        )
       )}
       {isSubCategorySelected && (
-        <>
-          {showOtherAddVenueBanner ? (
-            <FormLayout.Row>
-              <Callout
-                links={[
-                  {
-                    href: `/parcours-inscription/structure`,
-                    icon: {
-                      src: fullMoreIcon,
-                      alt: '',
-                    },
-                    label: 'Ajouter une structure',
-                  },
-                ]}
-                variant={CalloutVariant.ERROR}
-              >
-                Pour créer une offre, vous devez d’abord créer une structure.
-              </Callout>
-            </FormLayout.Row>
-          ) : (
-            <DetailsSubForm
-              isEanSearchDisplayed={isEanSearchDisplayed}
-              isProductBased={isProductBased}
-              isOfferCD={isSubCategoryCD(subcategoryId)}
-              readOnlyFields={readOnlyFields}
-              onImageUpload={onImageUpload}
-              onImageDelete={onImageDelete}
-              imageOffer={imageOffer}
-            />
-          )}
-        </>
+        <DetailsSubForm
+          isEanSearchDisplayed={isEanSearchDisplayed}
+          isProductBased={isProductBased}
+          isOfferCD={isSubCategoryCD(subcategoryId)}
+          readOnlyFields={readOnlyFields}
+          onImageUpload={onImageUpload}
+          onImageDelete={onImageDelete}
+          imageOffer={imageOffer}
+        />
       )}
     </>
   )
