--- pro/src/pages/IndividualOffer/IndividualOfferDetails/components/DetailsForm/SuggestedSubcategories/SuggestedSubcategories.module.scss
@@ -1,28 +1,0 @@
-@use "styles/mixins/_rem.scss" as rem;
-@use "styles/mixins/_size.scss" as size;
-
-.suggested-subcategories-section {
-  margin-bottom: 0;
-}
-
-.suggested-subcategories {
-  margin-bottom: rem.torem(24px);
-}
-
-.description {
-  margin-bottom: rem.torem(8px);
-}
-
-.items {
-  gap: rem.torem(16px);
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.suggested-subcategory {
-  width: 100%;
-
-  @media (min-width: size.$tablet) {
-    width: fit-content;
-  }
-}
