+++ pro/src/pages/IndividualOffer/commons/isRecordStore.ts
@@ -1,0 +1,9 @@
+import { VenueTypeCode } from 'apiClient/v1'
+
+export const isRecordStore = (
+  venues: { venueTypeCode: VenueTypeCode }[]
+): boolean => {
+  return venues.some(
+    (venue) => venue.venueTypeCode === ('RECORD_STORE' as VenueTypeCode)
+  )
+}
